+++ R/like.R
@@ -76,7 +76,11 @@ like <- function(x, pattern) {
   if (is.factor(x)) {
     as.integer(x) %in% base::grep(pattern, levels(x), ignore.case = TRUE)
   } else {
-    base::grepl(pattern, x, ignore.case = TRUE)
+    tryCatch(base::grepl(pattern, x, ignore.case = TRUE),
+             error = function(e) ifelse(test = grepl("Invalid regexp", e$message),
+                                        # try with perl = TRUE:
+                                        yes = return(base::grepl(pattern, x, ignore.case = TRUE, perl = TRUE)),
+                                        no = stop(e$message)))
   }
 }